Feat: add user-event.html
When HTML is loaded, it fetches information about the ongoing events from the server. Additionally, it retrieves events in which the current user is participating. The names of ongoing events are displayed on the screen, and if the user is participating in an event, it is indicated as 'Participating Event' next to the respective event. Clicking on an event name will navigate to the page for purchasing seats for that event.